+++ src/cli.py
@@ -3,6 +3,7 @@ import traceback
 import typing as t
 
 from handlers import (
+    check_for_updates_handler,
     fetch_handler,
     list_settings_handler,
     search_handler,
@@ -15,17 +16,13 @@ class CLIActions:
     FETCH = "fetch"
     SETTINGS_LIST = "settings_list"
     TOTP = "totp"
+    CHECK_FOR_UPDATES = "check_for_updates"
 
     @classmethod
     def choices(cls) -> t.List[str]:
         """Returns an action list."""
 
-        return [
-            cls.SEARCH,
-            cls.FETCH,
-            cls.SETTINGS_LIST,
-            cls.TOTP,
-        ]
+        return [cls.SEARCH, cls.FETCH, cls.SETTINGS_LIST, cls.TOTP, cls.CHECK_FOR_UPDATES]
 
 
 def parse_args() -> argparse.Namespace:
@@ -46,6 +43,9 @@ def parse_args() -> argparse.Namespace:
     totp_parser = subparsers.add_parser(CLIActions.TOTP)
     totp_parser.set_defaults(handler=totp_handler)
     totp_parser.add_argument("query")
+
+    check_for_updates_parser = subparsers.add_parser(CLIActions.CHECK_FOR_UPDATES)
+    check_for_updates_parser.set_defaults(handler=check_for_updates_handler)
 
     return parser.parse_args()